Added variable list and measurement history

+++ apps/nextjs/components/userVariable/measurements/measurement-delete-button.tsx
@@ -16,21 +16,18 @@ import {
 } from "@/components/ui/credenza"
 import { toast } from "@/components/ui/use-toast"
 import { Icons } from "@/components/icons"
+import { Measurement } from "@/app/types.ts";
 
 interface MeasurementsDeleteButtonProps {
-  measurements: {
-    id: string
-    date: Date
-    count: number
-    userVariable: {
-      id: string
-      name: string
-    }
-  }
+  measurement: Measurement
 }
 
-async function deleteUserVariable(userVariableId: string, measurementsId: string) {
-  const response = await fetch(`/api/userVariables/${userVariableId}/measurements/${measurementsId}`, {
+async function deleteMeasurement(measurementId: number | undefined) {
+  if(!measurementId) {
+    console.error("Measurement ID is not defined for deleteMeasurement")
+    return false
+  }
+  const response = await fetch(`/api/dfda/measurements?id=${measurementId}`, {
     method: "DELETE",
   })
 
@@ -49,14 +46,14 @@ async function deleteUserVariable(userVariableId: string, measurementsId: string
   return true
 }
 
-export function MeasurementsDeleteButton({ measurements }: MeasurementsDeleteButtonProps) {
+export function MeasurementDeleteButton({ measurement }: MeasurementsDeleteButtonProps) {
   const router = useRouter()
   const [showDeleteAlert, setShowDeleteAlert] = React.useState<boolean>(false)
   const [isDeleteLoading, setIsDeleteLoading] = React.useState<boolean>(false)
 
   const handleDelete = async () => {
     setIsDeleteLoading(true)
-    const deleted = await deleteUserVariable(measurements.userVariable.id, measurements.id)
+    const deleted = await deleteMeasurement(measurement.id)
 
     if (deleted) {
       setIsDeleteLoading(false)
@@ -80,7 +77,7 @@ export function MeasurementsDeleteButton({ measurements }: MeasurementsDeleteBut
         <CredenzaContent>
           <CredenzaHeader>
             <CredenzaTitle>
-              Delete measurements from {formatDate(measurements.date)}?
+              Delete measurement from {formatDate(measurement.startAt)}?
             </CredenzaTitle>
             <CredenzaDescription>
               This action cannot be undone.
